Cleaned unnecessary targets, renamed New_main.cpp to main.cpp and created 3D example with hanging node
